Refactor the way piece movesets are determined
Before, a simple boolean function was used to determine whether
a piece can move to a certain tile, but now every possible move
is generated as a piece is picked up, meaning the calculation
only needs to be done once. This will also eventually be useful
when creating an AI to play.